requirements.txt
fastapi 
uvicorn 
python-doctr
python-multipart
docx2pdf 
pdf2image
tqdm